feat(cache): server-side caching implemented

+++ src/services/postgres/NotesService.js
@@ -8,9 +8,10 @@ const NotFoundError = require("../../api/notes/exceptions/NotFoundError");
 const { mapDBToModel } = require("../../utils");
 
 class NotesService {
-	constructor(collaborationService) {
+	constructor(collaborationService, cacheService) {
 		this._pool = new Pool();
 		this._collaborationService = collaborationService;
+		this._cacheService = cacheService;
 	}
 
 	async verifyNoteAccess(noteId, userId) {
@@ -66,20 +67,39 @@ class NotesService {
 			throw new InvariantError("Catatan gagal ditambahkan");
 		}
 
+		await this._cacheService.delete(`notes:${owner}`);
+
 		return result.rows[0].id;
 	}
 
 	async getNotes(owner) {
-		const query = {
-			text: `SELECT notes.* FROM notes
+		try {
+			const cacheNotes = await this._cacheService.get(`notes:${owner}`);
+			return JSON.parse(cacheNotes);
+		} catch (error) {
+			const query = {
+				text: `SELECT notes.* FROM notes
 							LEFT JOIN collaborations ON collaborations.note_id = notes.id
 							WHERE notes.owner = $1 OR collaborations.user_id = $1
 							GROUP BY notes.id`,
-			values: [owner],
-		};
+				values: [owner],
+			};
 
-		const result = await this._pool.query(query);
-		return result.rows.map(mapDBToModel);
+			const result = await this._pool.query(query);
+
+			if (!result.rows.length) {
+				throw new NotFoundError("Catatan tidak ditemukan");
+			}
+
+			const mappedResult = result.rows.map(mapDBToModel);
+
+			await this._cacheService.set(
+				`notes:${owner}`,
+				JSON.stringify(mappedResult)
+			);
+
+			return mappedResult;
+		}
 	}
 
 	async getNoteById(id) {
@@ -100,7 +120,7 @@ class NotesService {
 		const updatedAt = new Date().toISOString();
 
 		const query = {
-			text: "UPDATE notes SET title = $1, body = $2, tags = $3, updated_at = $4 WHERE id = $5 RETURNING id",
+			text: "UPDATE notes SET title = $1, body = $2, tags = $3, updated_at = $4 WHERE id = $5 RETURNING id, owner",
 			values: [title, body, tags, updatedAt, id],
 		};
 
@@ -111,11 +131,14 @@ class NotesService {
 				"Gagal memperbarui catatan. Id tidak ditemukan"
 			);
 		}
+
+		const { owner } = result.rows[0];
+		await this._cacheService.delete(`notes:${owner}`);
 	}
 
 	async deleteNoteById(id) {
 		const query = {
-			text: "DELETE FROM notes WHERE id = $1 RETURNING id",
+			text: "DELETE FROM notes WHERE id = $1 RETURNING id, owner",
 			values: [id],
 		};
 
@@ -126,6 +149,8 @@ class NotesService {
 				"Gagal memperbarui catatan. Id tidak ditemukan"
 			);
 		}
+		const { owner } = result.rows[0];
+		await this._cacheService.delete(`notes:${owner}`);
 	}
 }
